% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_targets_project.R
\name{create_targets_project}
\alias{create_targets_project}
\title{This package will create a function called create_project()}
\usage{
create_targets_project(path, ...)
}
\arguments{
\item{path}{path to new project}

\item{...}{additional parameters}
}
\description{
It's callback is at: inst/rstudio/templates/project/create_targets_project.dcf
}
