% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_fns.R
\name{document_file}
\alias{document_file}
\title{Documentation wrapper that uses new roclet}
\usage{
document_file(my_file)
}
\arguments{
\item{my_file}{a roxygenized .R file}
}
\value{
a list of roxygen tags from the file. A side effect is that function documentation is produced using Quarto.
}
\description{
Documentation wrapper that uses new roclet
}
