% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper1_20230619.R
\name{helper1_20230619}
\alias{helper1_20230619}
\title{Helper for building a new target in a pipeline}
\usage{
helper1_20230619(my_params)
}
\arguments{
\item{my_params}{a list of parameters provided by the user via targetsAddin()}
}
\value{
a list of file names and file contents
}
\description{
Helper for building a new target in a pipeline
}
