% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper2_20230726.R
\name{helper2_20230726}
\alias{helper2_20230726}
\title{Helper for building a new target in a pipeline}
\usage{
helper2_20230726(my_params)
}
\arguments{
\item{my_params}{a list of parameters provided by the user via targetsAddin()}
}
\value{
a list of file names and file contents
}
\description{
Helper for building a new target in a pipeline
}
