% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_book.R
\name{render_book}
\alias{render_book}
\title{Convert roxygenized custom functions into a Quarto book. Works for projects that are not packages.}
\usage{
render_book(my_extension, my_fns)
}
\arguments{
\item{my_extension}{String indicating which Quarto extension to use}

\item{my_fns}{Character vector of roxygenized functions to document in a book}
}
\value{
Renders a Quarto book as a side effect. Returns the directory of the Quarto book.
}
\description{
Convert roxygenized custom functions into a Quarto book. Works for projects that are not packages.
}
