% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_fns.R
\name{roclet_output.roclet_memo}
\alias{roclet_output.roclet_memo}
\title{New roclet output}
\usage{
\method{roclet_output}{roclet_memo}(x, results, base_path, ...)
}
\arguments{
\item{x}{standard roxygen2 input}

\item{results}{standard roxygen2 input}

\item{base_path}{standard roxygen2 input}

\item{...}{additional inputs}
}
\value{
standard roxygen2 output
}
\description{
New roclet output
}
