% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_fns.R
\name{roclet_process.roclet_memo}
\alias{roclet_process.roclet_memo}
\title{New roclet process}
\usage{
\method{roclet_process}{roclet_memo}(x, blocks, env, base_path)
}
\arguments{
\item{x}{standard roxygen2 input}

\item{blocks}{standard roxygen2 input}

\item{env}{standard roxygen2 input}

\item{base_path}{standard roxygen2 input}
}
\value{
standard roxygen2 output
}
\description{
New roclet process
}
