% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_fns.R
\name{roxy_tag_parse.roxy_tag_memo}
\alias{roxy_tag_parse.roxy_tag_memo}
\title{New roxygen tag}
\usage{
roxy_tag_parse.roxy_tag_memo(x)
}
\arguments{
\item{x}{standard roxygen2 input}
}
\value{
standard roxygen2 output
}
\description{
New roxygen tag
}
