% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarchitect.R
\name{tarchitect}
\alias{tarchitect}
\title{Shiny app for building a new target in a pipeline}
\usage{
tarchitect(script = targets::tar_config_get("script"), code_folder = "R/")
}
\arguments{
\item{script}{Character of length 1, path to the target script file.
See \link[targets]{tar_manifest} for details.}

\item{code_folder}{a character vector of length 1; the name of a valid directory for R scripts.}
}
\value{
a list of file names and file contents.
The app has a side effect of creating a file
with the specified names and contents.
}
\description{
Shiny app for building a new target in a pipeline
}
