% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarchitect_exit.R
\name{tarchitect_exit}
\alias{tarchitect_exit}
\title{Exit tarchitect function}
\usage{
tarchitect_exit(exit_list, envir = parent.frame(n = 2))
}
\arguments{
\item{exit_list}{a list object; the output of shiny::runGadget()}
}
\description{
Exit tarchitect function
}
