% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetthis-package.R
\docType{package}
\name{targetthis-package}
\alias{targetthis}
\alias{targetthis-package}
\title{targetthis: Shiny widget to help build a `targets` pipeline}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Shiny widget to help build a `targets` pipeline.
}
\keyword{internal}
