% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_function_script.R
\name{write_function_script}
\alias{write_function_script}
\title{Helper for building a new target in a pipeline}
\usage{
write_function_script(new_function, code_folder = "R", text_fn)
}
\arguments{
\item{new_function}{a character vector of length 1}

\item{code_folder}{a character vector of length 1}

\item{text_fn}{a character vector of length 1; the function text}
}
\value{
a list of file names and file contents
}
\description{
Helper for building a new target in a pipeline
}
