% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper2_20230726.R
\name{helper2_20230726}
\alias{helper2_20230726}
\title{Helper for building a new target in a pipeline}
\usage{
helper2_20230726(new_inputs_df, new_target, new_function)
}
\arguments{
\item{new_inputs_df}{a data frame}

\item{new_target}{a character vector of length 1}

\item{new_function}{a character vector of length 1}
}
\value{
a list of file names and file contents
}
\description{
Helper for building a new target in a pipeline
}
