% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarchitect-package.R
\docType{package}
\name{tarchitect-package}
\alias{tarchitect-package}
\title{tarchitect: Shiny widget to help build a `targets` pipeline}
\description{
Shiny widget to help build a `targets` pipeline.
}
\keyword{internal}
