% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetthis-package.R
\docType{package}
\name{targetthis-package}
\alias{targetthis}
\alias{targetthis-package}
\title{targetthis: RStudio and `roxygen2` extensions for `targets` pipelines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extends RStudio and `roxygen2` to assist in building and documenting `targets` pipelines.
}
\keyword{internal}
